pyproject.toml
[project]
name = "yumbox"
version = "0.0.2"
description = "Tools and utilities for fast prototyping of machine learning ideas!"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Shayan", email = "shayaneshteharipour@gmail.com" }]
dependencies = [
    "numpy",
    "torch",
    "faiss-cpu",
    "hazm",
    "nltk",
    "unicodeblock",
    "lxml",
    "parsel",
    "tqdm",
    "pillow",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "flake8>=4.0", "black>=23.0", "mypy>=0.990"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["yumbox"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--strict-markers"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
